EA 23.335 Stable Patch 2

+++ Elin/Card.cs
@@ -4231,7 +4231,7 @@ public class Card : BaseCard, IReservable, ICardParent, IRenderSource, IGlobalVa
 			bonus = 0;
 			dmg = dmg * 100 / Mathf.Max(100 + num, 1);
 		}
-		int num4 = Dice.Roll(num2, sides, bonus, this);
+		long num4 = Dice.Roll(num2, sides, bonus, this);
 		dmg -= num4 * mod / 100;
 		if (dmg < 0)
 		{
